requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.4
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.31.0
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
gspread==6.1.2
google-api-python-client>=2.0.0
google-auth>=2.0.0
